Crate opa_util

Modules

serde

Serde

Structs

AccountSessionFields

AllOver

AnyOver

DeleteDocument

DeleteDocumentParams

DeletePolicy

GetConfig

GetConfigParams

GetDocument

GetDocumentParams

GetDocumentResult

GetPolicy

GetPolicyParams

GetStatus

GetStatusParams

Health

HealthParams

ListPolicies

OPAClient

OPAPolicyASTNodeRef

OPAPolicyASTNodeRefRef

OPAQuery

OPAQueryConfig

OPAQueryInput

OPAQueryResult

OPAResult

OPATxEntity

OPAType

PEQuery

PEQueryInput

PEQueryParams

PatchDocument

uses json patch format

Unknown

UpsertDocument

UpsertDocumentParams

UpsertPolicy

UpsertPolicyParams

_OPAClient

Enums

CowSlice

DisableAuthorization

use this type in places where authorization disabling needs to be very clear in all contexts e.g. as a function parameter, where both the function and the function caller should be unambiguously aware of what the parameter they are using / passing means

Evaluation

Filter

FilterVar

LogicalOp

OPAPolicyASTNode

OPAPolicyPathNode

OPAQueryInputAction

PEQueryInputAction

Constants

DEFAULT_TIMEOUT_SECONDS

Traits

AllOPAIdentifiable

AuthzService

AuthzServiceEntity

OPAContext

OPAIdentifiable

OPAIdentifiableEntity

OPAKey

OPATxCacheClient

OPATxCacheContext

Functions

pe_query

query_all

Type Definitions

OPADataCacheUpsert

Derive Macros

OPAContext

OPATxCacheContext

OPAType